# openapiv2/gen/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "prodvana/agent/agent_interaction.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/managed_resources": {
      "get": {
        "operationId": "ResourceManager_ListManagedResources",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/managed_resourceListManagedResourcesResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "prodvana.managed_resource.ResourceManager"
        ]
      }
    },
    "/v1/managed_resources/k8s/configure": {
      "post": {
        "operationId": "ResourceManager_ConfigureK8sCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/managed_resourceConfigureK8sClusterResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/managed_resourceConfigureK8sClusterReq"
            }
          }
        ],
        "tags": [
          "prodvana.managed_resource.ResourceManager"
        ]
      }
    },
...
  },
  "definitions": {
...
    "ConfigureK8sClusterReqEKSOptions": {
      "type": "object",
      "properties": {
        "accessKey": {
          "type": "string"
        },
        "secretKey": {
          "type": "string"
        },
        "accountId": {
          "type": "string"
        },
        "roleArn": {
          "type": "string"
        }
      }
    },
    "ConfigureK8sClusterReqGcpOptions": {
      "type": "object",
      "properties": {
        "project": {
          "type": "string"
        },
        "credentials": {
          "type": "string",
          "title": "TODO(naphat) there should be an integrations table for this and elevate to top level data model"
        }
      }
    },
...
    "ManagedResourceRun": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "terminal": {
          "type": "boolean"
        },
        "creationTimestamp": {
          "type": "string",
          "format": "date-time"
        },
        "lastUpdateTimestamp": {
          "type": "string",
          "format": "date-time"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "managed_resourceConfigureK8sClusterReq": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "gcpOptions": {
          "$ref": "#/definitions/ConfigureK8sClusterReqGcpOptions"
        },
        "eksOptions": {
          "$ref": "#/definitions/ConfigureK8sClusterReqEKSOptions"
        },
        "instanceType": {
          "type": "string"
        },
        "nodepoolZones": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "nodeCountPerZone": {
          "type": "integer",
          "format": "int32"
        },
        "region": {
          "type": "string"
        }
      }
    },
    "managed_resourceConfigureK8sClusterResp": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "runId": {
          "type": "string",
          "format": "date-time"
        },
        "version": {
          "type": "string"
        },
        "archiveUrl": {
          "type": "string",
          "title": "TODO(naphat) delete archive_url once we automate pulumi runs"
        }
      }
    },
...
    "managed_resourceListManagedResourcesResp": {
      "type": "object",
      "properties": {
        "managedResources": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/managed_resourceManagedResource"
          }
        }
      }
    },
    "managed_resourceManagedResource": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "latestRun": {
          "$ref": "#/definitions/ManagedResourceRun"
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
  }
}